Fixed failure of socket initialization on Solaris 10.
The patch is intended for compatibility with Solaris 10, which does not support the SO_RCVTIMEO and SO_SNDTIMEO options for sockets. The patch is intentionally made as simple as possible, without checking the OS version. Please note that these options are already present in Solaris 11 and the patch is not required for correct operation.